+++ src/app/(app)/layout.tsx
@@ -1,6 +1,12 @@
 import Footer from "@/components/layout/Footer";
 import Header from "@/components/layout/Header";
+import { defaultMetadata } from "@/utils/metaTag";
+import { Metadata } from "next";
 import { ReactNode } from "react";
+
+export const metadata: Metadata = {
+  ...defaultMetadata,
+};
 
 const layout = ({ children }: { children: ReactNode }) => {
   return (
