+++ src/app/(auth)/layout.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 import Image from "next/image";
 import Link from "next/link";
 import { ReactNode } from "react";
